Move end-to-end tests to their own file
